sim_tb: Fixup test bench to get it to run
Add a missing DBG_WISHBONE_SUPPORTED check, without it
we get failures when DBG_WISHBONE_SUPPORTED is not defined.

Add a $finish statement to allow the bench to exit.

Fix comments about the current jtag more we expect the state
machine to be in.

After this a basic test passes:

  fusesoc run --target sim_tb adv_debug_sys --iverilog_options "-D DBG_WISHBONE_SUPPORTED=1 -D ADBG_USE_HISPEED=1"

Non HISPEED test fails with CRC errors.

  fusesoc run --target sim_tb adv_debug_sys --iverilog_options "-D DBG_WISHBONE_SUPPORTED=1"

CPU0 test fails as the cpu_behavior model becomes hung.

  fusesoc run --target sim_tb adv_debug_sys --iverilog_options "-D DBG_WISHBONE_SUPPORTED=1 -D ADBG_USE_HISPEED=1 -D DBG_CPU0_SUPPORTED=1"